Benchmark and Search Modes improvements

+++ tools.hpp
@@ -56,10 +56,10 @@ inline void hexStrToBin(std::string str, uint8_t* data) {
 	for (uint16_t i(0) ; i < v.size() ; i++) data[i] = v[i];
 }
 
-inline uint32_t decodeCompact(uint32_t nCompact) {
-	const uint32_t nSize(nCompact >> 24), nWord(nCompact & 0x007fffff);
-	if (nSize <= 3) return nWord >> 8*(3 - nSize);
-	else return nWord << 8*(nSize - 3); // warning: this has problems if difficulty (uncompacted) ever goes past the 2^32 boundary
+inline uint64_t decodeCompact(uint32_t nCompact) { // Bitcoin Core's arith_uint256::SetCompact for UInt64_Ts
+	const uint64_t nSize(nCompact >> 24), nWord(nCompact & 0x007fffff);
+	if (nSize <= 3) return nWord >> (8ULL*(3ULL - nSize));
+	else return nWord << (8ULL*(nSize - 3ULL));
 }
 
 // Get address type (P2PKH, P2SH, Bech32), no proper support for Bech32 currently
